web forward feature
Ability to perform 301 redirects for some requests (in addition normal web content serving)
Motivation is that google sites cannot be associated with a naked domain (like example.com), only with www.example.com

A `host` field of incoming request (HTTP/1.1 only) is compared with list of hosts, defined by `--forward host url` option.
If a match is found, the request path is redirected to url+path
Otherwise, it is served as a request to the local file system.